requirements.txt
Django==1.11.5
django-debug-toolbar==1.8
django-extensions==1.9.1
pytz==2017.2
six==1.11.0
sqlparse==0.2.4
